### 2023-02-20 #5175 @editorialbot
[PRE REVIEW]: Geodata-Harvester: A Python package to jumpstart geospatial data extraction and analysis
**Submitting author:** @sebhaan (<a href="http://orcid.org/0000-0002-5994-5637">Sebastian Haan</a>)
**Repository:** https://github.com/Sydney-Informatics-Hub/geodata-harvester
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
