### 2023-02-22 738c380
Add missing DOIs and fix minor typos in paper
