+++ paper/paper.bib
@@ -1,4 +1,5 @@
 @article{Gorelick:2017,
+  doi={10.1016/j.rse.2017.06.031},
   title={Google Earth Engine: Planetary-scale geospatial analysis for everyone},
   author={Gorelick, Noel and Hancher, Matt and Dixon, Mike and Ilyushchenko, Simon and Thau, David and Moore, Rebecca},
   journal={Remote sensing of Environment},
@@ -9,6 +10,7 @@
 }
 
 @article{Jeffrey:2001,
+  doi={10.1016/S1364-8152(01)00008-1},  
   title={Using spatial interpolation to construct a comprehensive archive of Australian climate data},
   author={Jeffrey, Stephen J and Carter, John O and Moodie, Keith B and Beswick, Alan R},
   journal={Environmental Modelling \& Software},
