### 2023-02-22 #5175 @sebhaan
> Hi @sebhaan and thanks for your submission! Could you take a look at the missing DOIs above and see if they apply? They aren't always correct but should be checked.
>
> Also please take a look at the list of volunteer reviewers to suggest some relevant potential reviewers: https://bit.ly/joss-reviewers. It's also fine to suggest from your extended network, as long as you aren't conflicted. Please list github handles without "@" so they aren't preemptively pinged.

Thanks, I have checked the missing DOIs and added to paper.bib.
