+++ quarto/index.qmd
@@ -11,6 +11,8 @@ subtitle: "Jumpstart your analysis with a ready-made set of spatial-temporal ali
 
 The Geodata-Harvester project enables researchers with reusable workflows and provides open-source software for automatic data extraction from a wide range of data sources including spatial-temporal processing. User provided data is auto-completed with a suitable set of spatial- and temporal-aligned covariates as a ready-made dataset for machine learning models. All data layer maps are automatically extracted and aligned for a specific region and time period.
 
+{% embed_chat.html %}
+
 
 ## Data Sources
 
