### 2023-03-27 acaa7d2
change height of embed chatbot
