### 2023-04-08 #5205 @sjsrey
> @sjsrey & @martibosch : could you please do `@editorialbot generate my checklist` to make sure everything works?

It isn't clear to me how to proceed from the instructions?
