### 2023-04-11 #5205 @hugoledoux
@sjsrey you mean for the JOSS review or for the package you are reviewing?

If for JOSS, you need to generate the list of criteria we use, and then go through the list and check those that the submission fulfils. It's better explained there: https://github.com/openjournals/joss-reviews/issues/5205#issue-1606197484

If some parts are not fulfilled, then you write it here in this issue and you can provide a solution if you know how. We're here to help the submission be at the "JOSS level", to help the authors.
