### 2023-04-24 22fd00e
Add temporal processing tests for SILO
