### 2023-04-24 6625b03
add test function for multiband DEA time binning
