+++ .github/workflows/quarto-publish.yml
@@ -1,40 +1,0 @@
-on:
-  push:
-    branches: main
-
-name: Render and Publish
-
-jobs:
-  build-deploy:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        shell: bash -l {0}
-    steps:
-      - name: Check out repository
-        uses: actions/checkout@v2 
-
-      # - name: Create Python environment
-      #   uses: conda-incubator/setup-miniconda@v2
-      #   with:
-      #     auto-update-conda: false
-      #     python-version: 3.9    # set python version here
-      #     mamba-version: "*"
-      #     channels: conda-forge
-      #     channel-priority: true
-      #     activate-environment: quarto-import
-      #     environment-file: _import/environment.yml 
-        
-      - name: Set up Quartow
-        uses: quarto-dev/quarto-actions/setup@v2
-        with:
-          tinytex: true 
-          # r-version: 4.2.1
-
-      - name: Publish to GitHub Pages (and render)
-        uses: quarto-dev/quarto-actions/publish@v2
-        with:
-          target: gh-pages
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} 
-      
