### 2023-04-26 aea3954
Fix silo filenames and add optional rm
