### 2023-04-26 3a6c554
Add multi-band temporal aggregation
