+++ src/geodata_harvester/temporal.py
@@ -20,8 +20,9 @@ def combine_rasters_temporal(
     file_list, channel_name="band", attribute_name="long_name"
 ):
     """
-    Combines multiple tif files into single xarray object. Assumes files are in
-    temporal order, and additional channels contain sequential time step data.
+    Combines multiple tif files into single xarray object. 
+    Assumes additional channels contain sequential time step data. 
+    If multiple files in file_list, files must be in temporal order and same data type.
     Also assumes files are of the same shape (x,y,t).
 
     Example:
@@ -32,7 +33,7 @@ def combine_rasters_temporal(
 
     Parameters
     ----------
-    file_list : list of filename strings in date order to concatenate.
+    file_list : str or list of filename strings in date order to concatenate.
         Expected to be of the form "x,y" or "x,y,z1"
     channel_name : string of coordinate dimension to concatentate (band, time,
         etc). Check options with rioxarray.open_rasterio('filename').coords
@@ -46,6 +47,9 @@ def combine_rasters_temporal(
     """
     print("Concatenating", channel_name, "and", attribute_name, "over", file_list)
     # file_list = glob(os.path.join(data_dir, '*.tif'))
+
+    if type(file_list) == str:
+        file_list = [file_list]
 
     # Append all data/channels, collect metadata lists
     array_list = []
@@ -88,6 +92,44 @@ def combine_rasters_temporal(
     del xdr.attrs[attribute_name]
 
     return xdr
+
+
+def multiband_raster_to_xarray(file_list, date_list = None, mask_bandname = None):
+    """
+    Converts a stack of multiband raster with different dates to an xarray object.
+    
+    Parameters
+    ----------
+    file_list : list of filename strings in date order to concatenate.
+    date_list : list of dates in date order to concatenate. 
+        If None provided, the dates will be extracted from the file names.
+        This assumes that the date is given at the end of the file name after an underscore.
+    """
+    # Extract the dates from the file names if no date list is provided
+    if date_list is None:
+        date_list = get_date_after_last_underscore(file_list)
+
+    # Check if the lists have the same length
+    assert len(file_list) == len(date_list), "File list and date list must have the same length."
+    
+    # Create an empty list to store the DataArrays
+    data_arrays = []
+
+    for file, date in zip(file_list, date_list):
+        # Read the raster file using rioxarray
+        xds = rioxarray.open_rasterio(file)
+
+        # Assign the time coordinate
+        xds = xds.assign_coords({"time": pd.to_datetime(date)})
+        xds = xds.expand_dims("time")
+
+        # Append the DataArray to the list
+        data_arrays.append(xds)
+
+    # Concatenate the DataArrays along the time dimension
+    stacked_data = xr.concat(data_arrays, dim="time")
+
+    return stacked_data
 
 
 def temporal_crop(xdr, start_time, end_time):
@@ -219,44 +261,67 @@ def aggregate_temporal(xdr,
     return outfname_list, agg_list
 
 
-def group_by_custom_periods(xdr, periods: int, agg_range: int):
+def get_date_after_last_underscore(file_list):
+    result = []
+
+    for filename in file_list:
+        split_string = filename.rsplit('_', 1)  # Split the string from the right side, keeping only the last part
+
+        # Check if the string was split
+        if len(split_string) > 1:
+            last_part = split_string[-1]  # Get the part after the last "_"
+        else:
+            last_part = filename  # If the string didn't have any "_", return the original string
+
+        # Remove the file format ending
+        last_part = last_part.rsplit('.', 1)[0]
+
+        result.append(last_part)
+
+    return result
 
 
[…]
