### 2023-04-28 b3ea4aa
Add settings file for temporal processing test
