### 2023-04-28 97a8056
Add test for GEE temporal processing
