### 2023-05-02 e9e4547
Add temporal processing to SILO
