### 2023-05-02 f145453
Add if statement for temporal process
