### 2023-05-03 f45c86e
include temporal processing for DEA
