### 2023-05-03 15bcb3c
Add temporal binning for GEE
