### 2023-05-03 d96289e
Add automatic search for missing data and fill with nan
