### 2023-05-03 d53d251
Speed-up temporal aggregation
