### 2023-05-04 0e558c3
Add case-insensitive search for nodata value names
