### 2023-05-04 e0bc8eb
Fix GEE time binning and output names
