### 2023-05-05 51ddce6
Add check for datetime object
