### 2023-05-05 2d99d0a
Add check if multiple dates exist for layer
