### 2023-05-08 ef1b5cb
Add faster method and multiple band support for raster data query using xarray
