### 2023-05-08 8ae0754
Add new multi-band raster query function to harvest
