### 2023-05-08 2466252
Fix band attribute writing for version update in rioxarray
