### 2023-05-08 0a29f04
Fix bug in datafile writing
