### 2023-05-08 9bd638e
Remove geometry for csv export and cleanup
