### 2023-05-09 1b779f8
Add check for duplicate column names
