### 2023-05-09 a88e050
rename summary log file and add documentation
