### 2023-05-09 0727205
Check for duplicates
