+++ notebooks/settings/settings_temporal1.yaml
@@ -1,0 +1,33 @@
+# This settings file works for Google Earth Engine functionality on
+# the Data Harvester notebook.
+
+infile: data/example-site_llara.csv
+outpath: results_temporal1
+colname_lat: Lat
+colname_lng: Long
+
+# Bounding box: Left, Top, Right, Bottom
+target_bbox: [149.769345, -30.335861, 149.949173, -30.206271]
+# Resolution of data download in arcseconds
+target_res: 6.0
+#target_dates: 2021
+date_min : 2022-08-01
+time_intervals : 4
+date_max : 2022-12-01
+time_buffer : 
+
+target_sources:
+  DEA:
+  - ga_ls8c_ard_3
+  SILO:
+    daily_rain: sum
+  GEE:
+    preprocess:
+      collection: [LANDSAT/LC09/C02/T1_L2]
+      buffer: null
+      bound: null
+      mask_clouds: True
+      reduce:
+      spectral: NDVI
+    download:
+      bands: [NDVI]
