+++ README.md
@@ -19,33 +19,33 @@ The Geodata-Harvester Python package offers reusable and automated workflows for
 
 For the R-package wrapper of the Geodata-Harvester, please visit the [Github dataharvesteR project](https://github.com/Sydney-Informatics-Hub/dataharvester).
 
-## Content
+## 📚 Table of Contents
 
-- [Introduction](#introduction)
-- [Data Sources](#data-sources)
-- [Functionality](#functionality)
-- [Key Features](#key-features)
-- [Installation](#installation)
+- [Introduction](#🌐-introduction)
+- [Data Sources](#🌍-data-sources)
+- [Functionality](#⚙️-functionality)
+- [Key Features](#🌟-key-features)
+- [Installation](#🔧-installation)
     - [Conda or Mamba](#conda-or-mamba)
     - [PyPI](#pypi)
     - [Google Earth Engine extension](#google-earth-engine-extension)
     - [Local development](#local-development)
     - [Workshop Cloud Sandbox](#workshop-cloud-sandbox)
-- [How to get started](#how-to-get-started)
-- [How to add new data source modules](#how-to-add-new-data-source-modules)
-- [Code reference API](#code-reference-api)
-- [Contributions](#contributions)
-- [Attribution and Acknowledgments](#attribution-and-acknowledgments)
-- [License](#license)
+- [How to get started](#🚀-how-to-get-started)
+- [How to add new data source modules](#➕-how-to-add-new-data-source-modules)
+- [Code reference API](#📚-code-reference-api)
+- [Contributions](#🤝-contributions)
+- [Attribution and Acknowledgments](#👏-attribution-and-acknowledgments)
+- [License](#📄-license)
 
 
-## Introduction
+## 🌐 Introduction
 
 There is an enormous amount of national/global space-time data that is free and accessible. Examples are the numerous satellite platforms, weather, soil landscape grid of Australia. Many have a temporal dimension so for any point in Australia you can extract a time series of remote sensing and weather data and soil and terrain site variables. In the case of time series covariates there are a number of post-processing steps that a user can undertake to extract meaning, e.g. temporal means, aggregating in time. All of the above is a non-trivial task and a workflow where a user could enter a point (s) and get a tidy data frame of data cube variables would be a step towards people understanding its value and being able to jumpstart their analysis. This project will contribute processing tools for finding, extracting and converting these key data layers.
 
 Developed as part of the Agricultural Research Federation (AgReFed), Geodata-Harvester is an open-source software that allows users to jumpstart their analysis with a suitable set of spatial-temporal aligned raster maps and dataframes.
 
-## Data Sources
+## 🌍 Data Sources
 
 A detailed list of all available layers and their description can be found in [Data Overview](quarto/docs/Data_Overview.md).
 
@@ -69,7 +69,7 @@ The main goal of the Data Harvester is to enable researchers with reusable workf
 
 Geodata-Harvester is designed as a modular and maintainable project in the form of a multi-stage pipeline by providing explicit boundaries among tasks. To encourage interaction and experimentation with the pipeline, multiple frontend notebooks and use case scenarios are provided.
 
-## Key Features
+## 🌟 Key Features
 
 Below is a list of features available for the geodata-harvester package. Please check the project Github webpage and notebooks for examples, data selection, and other settings.
 
@@ -86,7 +86,7 @@ Below is a list of features available for the geodata-harvester package. Please 
 
 For more features, please see the [API reference documentation](https://sydney-informatics-hub.github.io/geodata-harvester/API/geodata_harvester/index.html).
 
-## Installation
+## 🔧 Installation
 
 Geodata-Harvester can be run on cloud-servers (e.g., in JupyterHub environment) or on your local machine. 
[…]
