+++ README.md
@@ -24,19 +24,19 @@ For the R-package wrapper of the Geodata-Harvester, please visit the [Github dat
 - [Introduction](#-introduction)
 - [Data Sources](#-data-sources)
 - [Functionality](#-functionality)
-- [Key Features](#🌟-key-features)
-- [Installation](#🔧-installation)
+- [Key Features](#-key-features)
+- [Installation](#-installation)
     - [Conda or Mamba](#conda-or-mamba)
     - [PyPI](#pypi)
     - [Google Earth Engine extension](#google-earth-engine-extension)
     - [Local development](#local-development)
     - [Workshop Cloud Sandbox](#workshop-cloud-sandbox)
-- [How to get started](#🚀-how-to-get-started)
-- [How to add new data source modules](#➕-how-to-add-new-data-source-modules)
-- [Code reference API](#📚-code-reference-api)
-- [Contributions](#🤝-contributions)
-- [Attribution and Acknowledgments](#👏-attribution-and-acknowledgments)
-- [License](#📄-license)
+- [How to get started](#-how-to-get-started)
+- [How to add new data source modules](#-how-to-add-new-data-source-modules)
+- [Code reference API](#-code-reference-api)
+- [Contributions](#-contributions)
+- [Attribution and Acknowledgments](#-attribution-and-acknowledgments)
+- [License](#-license)
 
 
 ## 💡 Introduction
