### 2023-05-19 ddddce5
Fix bug in xarray time binning
