### 2023-05-30 #5205 @lukasbeuster
Hi @sebhaan,

sorry for the extensive delay while processing the submission.

Let me first say that I'm excited about the piece of software. I do think it has the potential you subscribe to it to ease access to EO data for researchers and will help people to get up and running with their analyses much faster than usual!

That being said, I do have some feedback after testing around a little:

Installation and first use:

- The documentation should be clearer around the settings.yml file needed to run the first harvest. You could suggest that people clone the repo for their first use so they can execute the example notebooks and try out both widget and manual settings creation with ease.
- You explicitly mention GEE as optional, yet require it to run the example notebooks. This adds an additional layer of complexity that I would suggest you avoid. If you could showcase the GEE initialisation in a separate example that would be better. I ended up removing GEE manually from the respective settings.yml files after failing to initialise (will try again separately).

Harvesting:

I encountered a few hiccups while exploring:

- The example_harvest notebook results in an "UnboundLocalError" on my end, something to do with a local variable in the harvest function:

```
UnboundLocalError                         Traceback (most recent call last)
Cell In[3], line 1
----> 1 df = harvest.run(infname, preview = True, return_df = True)

File /opt/anaconda3/envs/geodata_harvester/lib/python3.11/site-packages/geodata_harvester/harvest.py:272, in run(path_to_config, log_name, preview, return_df)
    270     layer_titles += [layername]
    271     layer_list += [layername]
--> 272     agg_list += ['None']
    273     continue
    275 xdr = temporal.multiband_raster_to_xarray(files_layer)

UnboundLocalError: cannot access local variable 'agg_list' where it is not associated with a value
```
- harvest preview: The resulting plots appear a bit messy (e.g., overlapping titles/flipped). I'm not sure they currently serve any function other than showing that there are some results because values are not shown. I'd either clean/remove the preview or make the output more useful.
- optional: Some sources take quite some time to download. It would be nice to have a disclaimer when harvesting 'slower' sources to expect a longer download time (i.e. the SILO max/min temperature calls in the example notebooks took > 20 minutes on my end).

Widgets:
I have a few points regarding the widgets. They are a cool feature, but currently not very intuitive.

- It would be nice to have some more explanation within the widget, e.g. to clarify the input needs (csv/yml/?). This should include the bounding box (maybe there is a way to automatically derive the bbox from the input, but that'd be optional in my mind - clear documentation in the widget would be my priority).

- Automatically populates data sources in the settings.yml even if no data is requested (ticked in the corresponding tab). E.g., adds a GEE prompt even if no GEE data is needed resulting in a GEE initialisation. This should be resolved so that the widget populates with selected data sources only.

- Widget creates erroneous settings file: The widget uses the terminology ```settings.temp_intervals```. This results in an error when harvesting:
```---------------------------------------------------------------------------
Cell In[15], line 1
----> 1 df = harvest.run(fname_settings, return_df = True)

File /opt/anaconda3/envs/geodata_harvester/lib/python3.11/site-packages/geodata_harvester/harvest.py:108, in run(path_to_config, log_name, preview, return_df)
    104 # Temporal range
    105 # convert date strings to datetime objects
    106 date_diff = (datetime.strptime(settings.date_max, "%Y-%m-%d")
    107     - datetime.strptime(settings.date_min, "%Y-%m-%d")).days
--> 108 if settings.time_intervals is not None:
    109     period_days = date_diff // settings.time_intervals
    110     if period_days == 0:

AttributeError: 'types.SimpleNamespace' object has no attribute 'time_intervals'
```
The ```validate_settings``` function doesn't pick this up. Could it be that the harvest function input was changed from ```temp_intervals``` to ```time_intervals``` at some point? If you could please investigate and potentially fix the widget output and validate_settings function.

Paper:
- references aren't populating in the paper.md

That's my take for now. Looking forward to hearing your and the other reviewers' responses.

Best
Lukas
