### 2023-06-05 #5205 @martibosch
Regarding the checklist:

## General checks

- **Data sharing** and **Reproducibility**: I think it would be great to have an example notebook that can run without a GEE account. Otherwise, it should be clearly noted that a GEE account is required to use this library. I actually have run the notebook with a GEE account and got the following error on `example_harvest.ipynb`:

```
Starting the data harvester -----
ℹ Found the following 7 sources: ['DEA', 'DEM', 'Landscape', 'Radiometric', 'SILO', 'SLGA', 'GEE']

Downloading from API sources -----
⊙ Initialising Earth Engine... 2.9s
✔ Done

⌛ Downloading Google Earth Engine data...
Running preprocess() -----
ℹ Number of image(s) found: 1
⊙ Applying scale, offset and cloud masks... 1.2s
⊙ Calculating spectral indices: NDVI... 0.5s
✔ Preprocessing complete
Running download() -----
ℹ Band(s) selected: ['NDVI_median']
ℹ Setting scale to ~160.2m, converted from 6.0 arcsec at latitude -30.27
ℹ Setting download dir to results/ee
⚑ ee_LANDSAT_975fd1b8.tif already exists, skipping download
✔ Google Earth Engine download(s) complete
Running preprocess() -----
ℹ Number of image(s) found: 0
✖ Can't process zero images. Processing stopped

Cell In[3], line 1
----> 1 df = harvest.run(infname, preview = True, return_df = True)

File ~/mambaforge/envs/geodata-harvester/lib/python3.10/site-packages/geodata_harvester/harvest.py:201, in run(path_to_config, log_name, preview, return_df)
    199 # run eeharvest
    200 gee_outpath = os.path.join(settings.outpath,'ee')
--> 201 gee = eeharvester.auto(config=tmp_config, outpath=gee_outpath)
    202 if not isinstance(gee.filenames, list):
    203     # convert to list
    204     gee.filenames = [gee.filenames]

File ~/mambaforge/envs/geodata-harvester/lib/python3.10/site-packages/eeharvest/harvester.py:640, in auto(config, outpath)
    637 else:
    638     # download single collection
    639     img = collect(config=config)
--> 640     img.preprocess()
    641     if outpath is None:
    642         img.download()

File ~/mambaforge/envs/geodata-harvester/lib/python3.10/site-packages/eeharvest/harvester.py:269, in collect.preprocess(self, mask_clouds, mask_probability, reduce, spectral, clip, **kwargs)
    267 if count.getInfo() < 1:
    268     msg.err("Can't process zero images. Processing stopped")
--> 269     raise ValueError("No image to process, check your date range")
    271 # Cloud and shadow masking
    272 if mask_clouds:

ValueError: No image to process, check your date range
```

## Functionality

* **Functionality**: see the point above.

## Documentation

- **A statement of need**: I think it could be improved based on my first point of my review (in my other comment).
- **Example usage**: see point above about **Data sharing** and **Reproducibility**
- **Automated tests**: while I see tests in the tests folder, I do not see any instructions on how to run them. I believe that the authors should add them in the README or contribution guidelines. Additionally, they should consider automating the test execution with GitHub Actions (or a similar tool) to setup a git-ops CI workflow that requires tests to pass for each push/PR.

## Software paper

- **A statement of need**: same as the point above in the "Documentation" section.
- **State of the field**: again, in line with the first point of my review (in my other comment), I believe the authors should improve this part by commenting how geodata-harvester compares to other similar tools, e.g., intake.

That is it for my review. Looking forward to the authors responses. Best,
Martí
