### 2023-06-05 #5205 @martibosch
A final note which I forgot: when I install geodata-harvester from conda-forge, I get a `ModuleNotFoundError`, which I do not get when I install it using pip.
