### 2023-06-06 #5205 @sebhaan
Hi @martibosch and @lukasbeuster,
Just wanted to drop you both a quick note to say a huge thanks for your feedback on this software. You suggestions and bug reports are super helpful and have given me a roadmap for potential enhancements. I'm gearing up the next development cycle beginning next week and I'll be taking a close look at all of your suggestions, including a detailed response to your comments. I will reach out if I need some clarification on your points.
Thanks again, and I'll keep you posted on the progress.
