### 2023-06-13 a096c65
Add exception if no GEE image found for time interval and continue loop
