### 2023-06-13 5f9901e
Remove direct gdal dependency and add custom function for slope and aspect ratio calculation
