### 2023-06-13 5c721a8
Add display function for yaml settings file
