### 2023-06-13 f289414
Fix bug by assigning agg_list
