### 2023-06-14 e775ec6
Simplify basic settings plus description
