+++ notebooks/settings/settings_harvest.yaml
@@ -1,46 +1,41 @@
-# This settings file works for Google Earth Engine functionality on
-# the Data Harvester notebook.
+# Basic settings file for the Geodata-Harvester w/o GEE.
 
+# Path to the input file containing the site locations:
 infile: data/example-site_llara.csv
-outpath: results/
+
+# Path to output directory (will be created if it doesn't exist):
+outpath: results_example_harvest/
+
+# Column names for latitude and longitude in input file:
 colname_lat: Lat
 colname_lng: Long
 
-# Bounding box: Left, Top, Right, Bottom
+# Bounding box: Left (min Long), Bottom (min Lat), Right (max Long), Top (max Lat)
 target_bbox: [149.769345, -30.335861, 149.949173, -30.206271]
-# Reolution of data download in arcseconds
-target_res: 6.0
-#target_dates: 2021
-date_min : 2022-10-01
-time_intervals : 4
-date_max : 2022-11-30
-time_buffer : 7
 
+# Resolution of data download in arcseconds (1 arcsec ~ 30m)
+target_res: 6.0
+
+# Start date and end date of image collection
+date_min : 2022-10-01
+date_max : 2022-11-30
+# Number of time intervals to split the image collection into
+time_intervals : 2
+
+# List of data sources to download 
+# see https://github.com/Sydney-Informatics-Hub/geodata-harvester#-data-sources
 target_sources:
   DEA:
   - landsat_barest_earth
   - ga_ls_ard_3
   DEM:
   - DEM
-  Landscape: [Slope, Aspect, Relief_300m]
+  Landscape: [Slope, Aspect]
   Radiometric:
   - radmap2019_grid_dose_terr_awags_rad_2019
   - radmap2019_grid_dose_terr_filtered_awags_rad_2019
   SILO:
-    daily_rain: mean
-    max_temp: median
-    min_temp: median
-    monthly_rain: sum
+  - monthly_rain: sum
   SLGA:
-    Bulk_Density: [0-5cm]
-    Clay: [0-5cm]
-  GEE:
-    preprocess:
-      collection: LANDSAT/LC09/C02/T1_L2
-      buffer: null
-      bound: null
-      mask_clouds: True
-      reduce: median
-      spectral: NDVI
-    download:
-      bands: NDVI
+  - Bulk_Density: [0-5cm]
+  - Clay: [0-5cm]
