### 2023-06-14 bf7b26b
Fix dict settings
