### 2023-06-14 b36abaf
Add example notebook and settings file with GEE
