### 2023-06-14 57021ce
Change settings name temp_intervals to time_intervals
