### 2023-06-14 8fdb489
automatic bbox calculation and more field descriptions
