### 2023-06-14 debb190
Fix bbox calculation
