### 2023-06-14 f2c9086
Change SILO agg to one selection instead multi select
