### 2023-06-14 206df45
Update documentation and widget settings
