### 2023-06-14 e3c9dcc
Add bbox validation and generation
