+++ src/geodata_harvester/settingshandler.py
@@ -6,6 +6,7 @@ import json
 import os
 import datetime
 from types import SimpleNamespace
+import geopandas as gpd
 from IPython.display import display, JSON
 
 # Defaul settings yaml file
@@ -38,7 +39,34 @@ def display_settings(fname_settings, print_option = "json"):
         print(json.dumps(settings, indent=4, sort_keys=True, default=DateEncoder))
     else:
         print("print_option must be 'display' or 'json'")
-    
+
+
+def check_bbox(settings):
+    """
+    Check if bbox is valid. 
+    If no bbox is given, the bbox is calculated from the input points.
+
+    Input:
+        settings: settings namespace
+
+    Output:
+        settings: updated settings with bbox
+    """
+    if (settings.target_bbox is None) | (settings.target_bbox == ""):
+        gdfpoints = gpd.read_file(settings.infile)
+        longs = gdfpoints[settings.colname_lng].astype(float)
+        lats = gdfpoints[settings.colname_lat].astype(float)
+        if settings.target_bbox is None:
+            settings.target_bbox = (
+                min(longs) - 0.05,
+                min(lats) - 0.05,
+                max(longs) + 0.05,
+                max(lats) + 0.05,
+            )
+    assert len(settings.target_bbox) == 4, "There must be 4 values in bbox list"
+    assert settings.target_bbox[2] > settings.target_bbox[0], "Bounding box[0] must be smaller than box[2]"
+    assert settings.target_bbox[3] > settings.target_bbox[1], "Bounding box[1] must be smaller than box[3]"
+    return settings
 
 
 def main(fname_settings=_fname_settings, to_namespace=True):
@@ -55,5 +83,8 @@ def main(fname_settings=_fname_settings, to_namespace=True):
     # settings.variable_name rather than settings['variable_name'])
     if to_namespace:
         settings = SimpleNamespace(**settings)
+
+    # Check if bbox is valid
+    settings = check_bbox(settings)
 
     return settings
