### 2023-06-15 fa346bd
Fix bug in aggregate_rasters function
