### 2023-06-15 cacd041
Add new notebook for step-by-step demonstration of processes
