+++ notebooks/settings/settings_simple.yaml
@@ -1,0 +1,50 @@
+# Basic settings file for the Geodata-Harvester w/o GEE.
+
+# Path to the input file containing the site locations:
+infile: data/example-site_llara.csv
+
+# Path to output directory (will be created if it doesn't exist):
+outpath: results_technical/
+
+# Column names for latitude and longitude in input file:
+colname_lat: Lat
+colname_lng: Long
+
+# Bounding box: Left (min Long), Bottom (min Lat), Right (max Long), Top (max Lat)
+target_bbox:
+
+# Resolution of data download in arcseconds (1 arcsec ~ 30m)
+target_res: 6.0
+
+# Start date and end date of image collection
+date_min : 2022-10-01
+date_max : 2022-11-30
+# Number of time intervals to split the image collection into
+time_intervals : 1
+
+# List of data sources to download 
+# see https://github.com/Sydney-Informatics-Hub/geodata-harvester#-data-sources
+target_sources:
+  DEA:
+  - landsat_barest_earth
+  DEM:
+  - DEM
+  Landscape: [Slope, Aspect]
+  Radiometric:
+  - radmap2019_grid_dose_terr_awags_rad_2019
+  SILO:
+    min_temp: mean
+    max_temp: mean
+    monthly_rain: sum
+  SLGA:
+    Bulk_Density: [0-5cm]
+  GEE:
+    preprocess:
+      collection: LANDSAT/LC09/C02/T1_L2
+      buffer: null
+      bound: null
+      mask_clouds: True
+      reduce: median
+      spectral: NDVI
+    download:
+      bands: NDVI
