### 2023-06-20 07bc8e5
Update test_getdata_dea
