### 2023-06-20 a99e2e2
Move test functions to tests and fix outfname in get_dea_images()
