+++ src/geodata_harvester/getdata_dea.py
@@ -787,8 +787,8 @@ def get_dea_images(
                 date[:-1]).astimezone(timezone.utc)
             fname_out = f"{layername}_{datestring.year}-{datestring.month}-{datestring.day}{fname_end}"
         outfname = os.path.join(outpath, fname_out)
-        if os.path.exists(outfname):
-            outfnames.append(outfname)
+        #if os.path.exists(outfname):
+        #    outfnames.append(outfname)
         # Get data
         download_ok = get_wcsmap(
             outfname,
@@ -802,8 +802,7 @@ def get_dea_images(
         )
         # Log download success message if file does not already exist
         if download_ok:
-            # outfnames.append(outfname)
-            pass
+            outfnames.append(outfname)
         else:
             cprint(f"✘ {layername} for date {date} failed to download", "red")
     # return [item for sublist in outfnames for item in sublist]
@@ -912,113 +911,3 @@ def get_dea_images_daterange(
         else:
             cprint(f"✘ {layername} for date {date} failed to download", "red")
     return outfnames
-
-
-### Some test functions ###
-
-
-def test_get_capabilities():
-    """
-    Test script to retrieve WCS capabilities
-    """
-    # DEA layer
-    url = "https://ows.dea.ga.gov.au/?version=1.3.0"
-    # Get capabilities
-    keys, title_list, description_list, bbox_list, timelimits, nbands = get_capabilities(url)
-    assert len(keys) > 0
-    print('Test passed')
-
-
-def test_get_times():
-    """
-    Test script to retrieve available times for a layer
-    """
-    layername = "ga_ls8c_ard_3"
-    url = "https://ows.dea.ga.gov.au/?version=1.3.0"
-    times = get_times(url, layername)
-    assert len(times) > 0
-    print('Test passed')
-
-
-def test_get_times_startend():
-    """
-    Test script to retrieve available times for a layer
-    """
-    layername = "ga_ls8c_ard_3"
-    url = "https://ows.dea.ga.gov.au/?version=1.3.0"
-    dt_start = "2020-01-01"
-    dt_end = "2020-02-20"
-    times = get_times_startend(url, layername, dt_start, dt_end)
-    assert len(times) > 0
-    print('Test passed')
-
-
-def test_get_wcsmap():
-    """
-    Test script to retrieve and save image for one layer and date
-    """
-    url = "https://ows.dea.ga.gov.au/?version=1.3.0"
-    layername = "ga_ls8c_ard_3"
-    times = get_times(url, layername)
-    crs = "EPSG:4326"  # WGS84
-    # define bounding box for retrieval (simple test here for entire Australia)
-    bbox = [130, -40, 150, -20]
-    # define resolution (in arcsecs per pixel since crs is in WGS84)
-    resolution = 100
-    # get latest image
-    time = times[-1]
-    # define output file name
-    fname_out = f"test_{layername}_{time}.tif"
-    # Get data
-    download_ok = get_wcsmap(fname_out, layername, bbox, time, resolution, url)
-    assert download_ok
-    print('Test passed')
-
-
-def test_get_dea_images():
-    """
-    Test script to retrieve and save images for a given year
-    """
-    url = "https://ows.dea.ga.gov.au/?version=1.3.0"
-    # Get data (here only for first layer)
-    layername = "ga_ls8c_ard_3"
-    # Crs for output
-    crs = "EPSG:4326"  # WGS84
-    # define bounding box for retrieval (simple test here for entire Australia)
-    bbox = [120, -40, 140, -20]
-    # define resolution (in arcsecs per pixel since crs is in WGS84)
-    resolution = 100
-    # define year
-    year = 2019
-    # define outpath
[…]
