### 2023-06-20 6fc4084
Speed-up automatic test
