+++ tests/test_getdata_radiometric.py
@@ -7,13 +7,13 @@ from geodata_harvester import getdata_radiometric
 
 
 def test_get_capabilities():
-    """
-    Test script to retrieve WCS capabilities
-    """
-    # Get capabilities
-    keys, titles, descriptions, bboxs = getdata_radiometric.get_capabilities()
-    assert len(keys) > 0
-    print("get_capabilities test passed")
+   """
+   Test script to retrieve WCS capabilities
+   """
+   # Get capabilities
+   keys, titles, descriptions, bboxs = getdata_radiometric.get_capabilities()
+   assert len(keys) > 0
+   print("get_capabilities test passed")
 
 
 def test_get_times():
@@ -26,19 +26,19 @@ def test_get_times():
     assert len(times) > 0
     print('get_times test passed')
 
-
-def test_times():
-    """
-    Check that there is only one time per layers.
-    """
-    url = "https://gsky.nci.org.au/ows/national_geophysical_compilations?service=WCS"
-    radiometricdict = getdata_radiometric.get_radiometricdict()
-    layernames = radiometricdict["layernames"]
-    for key in layernames:
-        times = getdata_radiometric.get_times(url, key)
-        print(f"{key}: {times}")
-        assert len(times) == 1
-    print("test_times passed")
+""" This test function takes a long time and is not necessary for download testing"""
+# def test_times():
+#     """
+#     Check that there is only one time per layers.
+#     """
+#     url = "https://gsky.nci.org.au/ows/national_geophysical_compilations?service=WCS"
+#     radiometricdict = getdata_radiometric.get_radiometricdict()
+#     layernames = radiometricdict["layernames"]
+#     for key in layernames:
+#         times = getdata_radiometric.get_times(url, key)
+#         print(f"{key}: {times}")
+#         assert len(times) == 1
+#     print("test_times passed")
 
 
 def test_get_radiometric_image():
@@ -49,8 +49,8 @@ def test_get_radiometric_image():
     # "radmap2019_grid_dose_terr_awags_rad_2019"  # for some layers readout time of 30s is exceeding (server limited)
     layername = "radmap2019_grid_dose_terr_filtered_awags_rad_2019"
     crs = "EPSG:4326"  # WGS84
-    # define bounding box for retrieval (simple test here for entire Australia)
-    bbox = (114, -44, 153.9, -11)
+    # define sall bounding box for retrieval test
+    bbox = (130, -30, 131, -29)
     # define resolution (in arcsecs per pixel since crs is in WGS84)
     resolution = 100
     # define output file name
