+++ src/geodata_harvester/getdata_radiometric.py
@@ -284,58 +284,3 @@ def get_times(url, layername, year=None):
             if datetime.fromisoformat(time[:-1]).astimezone(timezone.utc).year == year:
                 dates.append(time)
         return dates
-
-
-### Some test functions ###
-
-
-def test_get_capabilities():
-    """
-    Test script to retrieve WCS capabilities
-    """
-    # Get capabilities
-    keys, titles, descriptions, bboxs = get_capabilities()
-    assert len(keys) > 0
-
-
-def test_get_times():
-    """
-    Test script to retrieve available times for a layer
-    """
-    url = "https://gsky.nci.org.au/ows/national_geophysical_compilations?service=WCS"
-    times = get_times(url, layername)
-    assert len(times) > 0
-
-
-def test_times():
-    """
-    Check that there is only one time per layers.
-    """
-    url = "https://gsky.nci.org.au/ows/national_geophysical_compilations?service=WCS"
-    radiometricdict = get_radiometricdict()
-    layernames = radiometricdict["layernames"]
-    for key in layernames:
-        times = get_times(url, key)
-        print(f"{key}: {times}")
-        assert len(times) == 1
-
-
-def test_get_radiometric_image():
-    """
-    Test script to retrieve and save image for one layer
-    """
-    url = "https://gsky.nci.org.au/ows/national_geophysical_compilations?service=WCS"
-    # "radmap2019_grid_dose_terr_awags_rad_2019"  # for some layers readout time of 30s is exceeding (server limited)
-    layername = "radmap2019_grid_dose_terr_filtered_awags_rad_2019"
-    crs = "EPSG:4326"  # WGS84
-    # define bounding box for retrieval (simple test here for entire Australia)
-    bbox = (114, -44, 153.9, -11)
-    # define resolution (in arcsecs per pixel since crs is in WGS84)
-    resolution = 100
-    # define output file name
-    fname_out = f"test_{layername}.tif"
-    # Get data
-    download_ok = get_radiometric_image(
-        fname_out, layername, bbox, url, resolution=resolution
-    )
-    assert download_ok
