### 2023-06-20 f2092a5
Speed-up getdata_silo test
