### 2023-06-20 c514555
Include fallback for no time slice available
