### 2023-06-20 74e8679
Update test_harvest, no GEE required anymore
