+++ tests/settings/settings_test_harvest.yaml
@@ -2,38 +2,38 @@
 # the Data Harvester notebook.
 
 infile: data/example-site_llara.csv
-outpath: test_results_temp/
+outpath: test_results_harvest/
 colname_lat: Lat
 colname_lng: Long
 
 # Bounding box: Left, Top, Right, Bottom
-target_bbox: [149.769345, -30.335861, 149.949173, -30.206271]
+target_bbox: [149, -31, 150, -30]
 # Reolution of data download in arcseconds
-target_res: 6.0
+target_res: 100.0
 date_min : 2022-10-01
 time_intervals : 2
 date_max : 2022-11-01
-time_buffer : 2
 
 target_sources:
   DEA:
   - landsat_barest_earth
   DEM:
   - DEM
+  - Aspect
   Landscape: [Slope]
   Radiometric:
   - radmap2019_grid_dose_terr_awags_rad_2019
   SILO:
-    monthly_rain: sum
+    daily_rain: sum
   SLGA:
     Bulk_Density: [0-5cm]
-  GEE:
-    preprocess:
-      collection: LANDSAT/LC09/C02/T1_L2
-      buffer: null
-      bound: null
-      mask_clouds: True
-      reduce: median
-      spectral: NDVI
-    download:
-      bands: NDVI
+  # GEE:
+  #   preprocess:
+  #     collection: LANDSAT/LC09/C02/T1_L2
+  #     buffer: null
+  #     bound: null
+  #     mask_clouds: True
+  #     reduce: median
+  #     spectral: NDVI
+  #   download:
+  #     bands: NDVI
