+++ quarto/API_md/geodata_harvester/harvest.md
@@ -1,19 +1,26 @@
 Module geodata_harvester.harvest
 ================================
-This script is running the headless version of the data harvester.
+This script is running the headless version of the geodata-harvester.
 
 The following main steps are automatically executed within the run() function:
     - loading settings from config file
     - creating bounding box from input file points if not provided
-    - downloading data layers as sepcified in config file
+    - downloading data layers as specified in config file
+    - processing data layers as specified in config file
+    - save downloaded image files to disk as GeoTiffs
+    - save summary table of downloaded files as CSV
     - extract data for point locations provided in input file (name specified in settings)
-    - save results to disk as csv and geopackage
+    - save extracted point results to disk as CSV and as geopackage 
+
+Example call within Python:
+    from geodata_harvester import harvest
+    harvest.run(path_to_config))
 
 Functions
 ---------
 
     
-`run(path_to_config, log_name='download_log', preview=False, return_df=False)`
+`run(path_to_config, log_name='download_summary', preview=False, return_df=False)`
 :   A headless version of the Data-Harvester (with some limitations).
     Results are saved to disk.
     
