+++ README.md
@@ -31,6 +31,7 @@ For the R-package wrapper of the Geodata-Harvester, please visit the [Github dat
     - [Google Earth Engine extension](#google-earth-engine-extension)
     - [Local development](#local-development)
     - [Workshop Cloud Sandbox](#workshop-cloud-sandbox)
+- [Settings Overview](#settings-overview)
 - [How to get started](#-how-to-get-started)
 - [How to add new data source modules](#-how-to-add-new-data-source-modules)
 - [Code reference API](#-code-reference-api)
@@ -71,20 +72,18 @@ Geodata-Harvester is designed as a modular and maintainable project in the form 
 
 ## 🌟 Key Features
 
-Below is a list of features available for the geodata-harvester package. Please check the project Github webpage and notebooks for examples, data selection, and other settings.
+The geodata-harvester package provides the following core features:
 
-- enabling reproducible workflows via YAML settings files.
-- automatic data retrieval from geodata APIs for given locations and dates.
-- automatic download and spatiotemporal processing of geo-spatial maps for user-specified bounding box, resolution, and time-scale.
-- support for time-series data extraction for multiple time slices. 
-- automatic extraction of retrieved data into ready-made dataframes for ML training.
-- automatic generation of ready-made aligned maps in GeoTiff format.
-- preview of downloaded and aligned maps.
-- support for saving and loading settings via interactive widgets.
-- with connectivity support to the Google Earth Engine API, perform petabyte-scale operations which include temporal cloud/shadow masking and automatic calculation of spectral indices.
-- example notebooks and use-cases are provided for the user to get started.
+- enabling reproducible workflows via YAML settings files (see for settings example [settings_harvest.yaml](notebooks/settings/settings_harvest.yaml)).
+- interactive widgets for settings selection (see for example notebook [example_harvest_with_widgets.ipynb](https://github.com/Sydney-Informatics-Hub/geodata-harvester/tree/main/notebooks/example_harvest_with_widgets.ipynb)).
+- automated download and processing pipeline for multiple data sources (supported by the `harvest.run()` function as demonstrated in the example notebook [example_harvest.ipynb](https://github.com/Sydney-Informatics-Hub/geodata-harvester/tree/main/notebooks/example_harvest.ipynb)).
+- automatic data retrieval from geodata APIs and spatiotemporal processing for given locations, bounding box, resoultion, and time-scales into ready-made aligned and geo-referenced maps in GeoTiff format (see [notebook API download and process step](https://nbviewer.jupyter.org/github/Sydney-Informatics-Hub/geodata-harvester/blob/main/notebooks/example_harvest_stepwise.ipynb#Download-and-process-data-from-API-sources)).
+- support for time-series data extraction for multiple time slices (see example notebook [example_harvest_temporal1.ipynb](https://github.com/Sydney-Informatics-Hub/geodata-harvester/tree/main/notebooks/example_harvest_temporal1.ipynb) and [example_harvest_temporal2.ipynb](https://github.com/Sydney-Informatics-Hub/geodata-harvester/tree/main/notebooks/example_harvest_temporal2.ipynb)). 
+- automatic extraction of retrieved data into ready-made dataframes for ML training ([see notebook point extraction process](https://nbviewer.jupyter.org/github/Sydney-Informatics-Hub/geodata-harvester/blob/main/notebooks/example_harvest_stepwise.ipynb#Points-extraction-from-downloaded/processed-data))).
[…]
