### 2023-06-21 89bd983
Remove GEE settings if not selected in widget
