### 2023-06-22 #5205 @sebhaan
## Reply to reviewers

Dear @lukasbeuster and @martibosch,

thank you very much for your time and effort in reviewing our software and paper. We greatly appreciate your expert insight and constructive suggestions, which significantly improved both the usability of the Geodata-Harvester and the clarity of the documentation.

Your suggestions have also concluded in the release of the latest Geodata-Harvester v1.1.0 (see for release notes: https://github.com/Sydney-Informatics-Hub/geodata-harvester/releases/tag/v1.1.0)

We have addressed your comments and suggestions in the following. Please let me know if you have any further questions or suggestions.

Once again, many thanks for your valuable contribution to this process.

--------------

@lukasbeuster
## Reply to review by Lukas Beuster

### Installation and first use
- The documentation should be clearer around the settings.yml file needed to run the first harvest. You could suggest that people clone the repo for their first use so they can execute the example notebooks and try out both widget and manual settings creation with ease.
Reply: We have streamlined and added more documentation to the settings files. We have also added more documentation to the README file section "How to get started", including the step to clone the repo first.

- You explicitly mention GEE as optional, yet require it to run the example notebooks. This adds an additional layer of complexity that I would suggest you avoid. If you could showcase the GEE initialisation in a separate example that would be better. I ended up removing GEE manually from the respective settings.yml files after failing to initialise (will try again separately).
Reply: To simplify the use for the Geodata-Harvester, we have removed the GEE initialisation from the basic notebook example (example_harvest.ipynb/settings_harvest.yaml). In addition we have added a new notebook (example_harvest_withGEE.ipynb) to demonstrate GEE including authentication check. A new step-by-step notebook (example_harvest_stepwise.ipynb) has also been added, which includes an optional walk-through of the GEE initialisation and processing.

### Harvesting:
- The example_harvest notebook results in an "UnboundLocalError" on my end, something to do with a local variable in the harvest function
Reply: This error has been fixed and included in the latest release v1.1.0.
- harvest preview: The resulting plots appear a bit messy (e.g., overlapping titles/flipped). I'm not sure they currently serve any function other than showing that there are some results because values are not shown. I'd either clean/remove the preview or make the output more useful.
Reply: The preview plots are not meant for analysis but rather to quickly verify that the data has been downloaded and processed. If the preview is not needed, the user can easily set the argument `preview = False` in the harvest.run function. To analyse and use the data, the user need to load the resulting dataframes and maps as saved in the output folder with their favorite GIS software or python package.

- optional: Some sources take quite some time to download. It would be nice to have a disclaimer when harvesting 'slower' sources to expect a longer download time (i.e. the SILO max/min temperature calls in the example notebooks took > 20 minutes on my end).
Reply: We have added a warning for SILO data download time (commit d9feb0cd0f7318474f9f562369c92d4737bd8bb9) given that the SILO raster data API does not provide a cut-out with bounding box and the entire raster is downloaded and then processed. While it would be helpful to have an estimate for the time of other sources, there are currently many factors that influence the download time (e.g. internet speed, server load, layer types, selected time range, resolution, bbox) and we have not found a way to reliably estimate the download time. We have therefore decided to not include a time estimate for other sources.

### Widgets:
- It would be nice to have some more explanation within the widget, e.g. to clarify the input needs (csv/yml/?). This should include the bounding box (maybe there is a way to automatically derive the bbox from the input, but that'd be optional in my mind - clear documentation in the widget would be my priority).
Reply: We have added more descriptions for the setting widget fields and also updated the notebook example_harvest_with_widgets.ipynb for widget demonstration (commit 206df453e6685a09e80785d11c5f45c690596268). We also added an automatic inference of the bounding box if kept empty, based on the extent of the input data plus padding (commit 57021ceb2ba0aabbf757efaec284a8f07c47de8c and 8fdb48983553260d38c7d899bc90bd2b415bd015).

- Automatically populates data sources in the settings.yml even if no data is requested (ticked in the corresponding tab). E.g., adds a GEE prompt even if no GEE data is needed resulting in a GEE initialisation. This should be resolved so that the widget populates with selected data sources only.
Reply: We have fixed this issue in the notebook and removed the GEE settings if not selected (commit 89bd983011f7091354e604c88869e872b13687ac). We have also added a check for GEE initialisation in the notebook example_harvest_withGEE.ipynb.

- Widget creates erroneous settings file: The widget uses the terminology settings.temp_intervals. This results in an error when harvesting
Reply: Fixed (commit 57021ceb2ba0aabbf757efaec284a8f07c47de8c)

### Paper:
- references aren't populating in the paper.md
Reply: The references are saved in the paper.bib file and are automatically populated in the paper.pdf when the paper is rendered, e.g, with pandoc --citeproc (see paper.pdf on github). We have also checked the JOSS proof pdf (https://github.com/openjournals/joss-papers/blob/joss.05205/joss.05205/10.21105.joss.05205.pdf) and the references are correctly populated there as well. Could you please let us know if you are still experiencing this issue?

----------------------------

@martibosch
## Reply to review by Marti Bosch

- If I understand its prupose correctly, it appears very similar to intake - maybe a bit more tailored to geospatial (and temporal) data. Since it seems that the library tries to standardize geospatial data access, why not using the existing intake approach and write the implemented data sources as intake plugins? There may actually be a good reason not to do so, but in such a case, it would be interesting to explain it (maybe in the statement of need?), e.g., what are the advantages when compared to intake.
Reply: While I'm not very familiar with the intake package plugins, it seems like these plugins focus primarily on data loading/searching, perhaps not unlike other geodata loaders such as rioxarray, pystac and osgeo libraries for accessing and searching web APIs. In fact the geodata-harvester heavily utilises some of these geodata API handlers such as owslib and rioxarray to retrieve automatically API capabilities and to download/process data. Hence the Geodata-Harvester can be looked at as a high level interface/workflow to simplify automatic download and processing steps all at once for the user. In my opinion the most similar packages to Geodata-Harvester might be some of the open-source Python extension for Google Earth Engine, which are actually combined and applied for the geodata-harvester GEE connectivity. A perhaps alternative route might be to extend the Geodata-Harvester with high-level STAC catalogue retrieval functions (which could be added as extra module in principle).
We have expanded in the paper the Statement of Need and also added a comparison to other packages in the README file (see updated paper).

- after going through the paper, documentation and notebooks is that the scope is too broad so the provided features feel too abstract to the reader. I think it would help if the README included a simple snippet and pointed to an example notebook with a more specific use-case that provides clearer insights into the main features of the package. Currently, the notebooks barely include the harvest.run method and its resulting logs, but more text should be added so that the reader/potential user understands better what is going on and thus how the package may be useful. After all, mixing markdown and Python code is what makes Jupyter notebooks so powerful. Additionally, I would showing what a config file, e.g., settings_harvest.yaml, looks like within the notebook, e.g., see how %%writefile is used in cell 3 of the intake tutorial. I would first focus on explaining the automatic retrieval and basic spatiotemporal processing features (e.g., bounding box, spatial/temporal resampling...), then move on to describing the automatic generation of data frames and maps for ML (with the current notebooks, I actually do not understand how to use such a feature).

Reply: We have significantly improved the documentation as outline in the following:
1) we have added a detailed description in the notebooks for the harvest.run function that provides a list of all steps executed by the function.
2) we have added a new notebook (example_harvest_stepwise.ipynb) that goes through all the download/process steps one by one with detailed descriptions.
3) we updated the Key feature section in the README file and have added to each feature a link to an example code snippet in the respective notebooks that demonstrates this features.
4) we have added to the notebooks a section to display the settings file that is loaded before the harvest.run function is called.

- I have not been able to execute the notebooks since I do not have access to GEE - I requested access some weeks ago but have not heard back from them. Is it possible to use any of the features without a GEE account? If so, would it be possible to add a corresponding example notebook?

Reply: We have split the example_harvest.ipynb notebook into two, one that works w/o GEE (example_harvest.ipynb) and one including GEE (example_harvest_withGEE.ipynb), so GEE authentication is now optional. Also we have removed the GEE initialisation from the widget notebook if GEE is not selected. However we hope that your GEE account has been approved by now, typically it takes a few minutes. If not, please let us know and we can look into this issue.

- Given the current materials (notebooks, documentation...), as an external reader/user with some expertise in the field, I understand the features that you aim to provide but I have a hard time understanding how I may use this library within my work. In my opinion, this can be improved by providing more detailed (and concise) example notebooks. The SIH-AgReFed Workshop materials give a more thorough overview but may be overwhelming for someone who is just checking out the library.

Reply: Indeed the package includes a lot of functionality and modules for a wide range of APIs and processing. As outlined in the paper and README, the main idea of this project is to make it easier for researchers, or anybody working with GIS data, to find and extract key foundational datasets (such as climate, landscape, soil, and remote sensed data) at once given the required spatial, area and temporal range for their analysis. We hope the additional new notebooks,and updated documentation will help to clarify the use-cases for this package.
To showcase different setting options, we provide a range of templates for settings file in conjunction with the notebooks, and also created the widget notebook to simplify the setting selection. So far we have used this package for several projects and found that it is very useful for quickly downloading and processing data for a range of use cases such as feature aggregation for geospatial machine learning and agricultural research.

### General checks
- Data sharing and Reproducibility: I think it would be great to have an example notebook that can run without a GEE account. Otherwise, it should be clearly noted that a GEE account is required to use this library. I actually have run the notebook with a GEE account and got the following error on example_harvest.ipynb
ValueError: No image to process, check your date range

Reply: As mentioned above we removed the GEE selection from the starter notebook example_harvest.ipynb, and also updated the widget notebook to remove the GEE initialisation if GEE is not selected. Where GEE is required we added clear notes that a GEE account is needed and guide the user to the GEE account instructions.
The error has been fixed (commit a096c657b8d194461b043cc0115340a2e1d26da6).

### Documentation
- A statement of need: I think it could be improved based on my first point of my review (in my other comment).
Reply: As mentioned above we have expanded on the Statement of Need in the paper.

- Example usage: see point above about Data sharing and Reproducibility
Reply: See reply above

- Automated tests: while I see tests in the tests folder, I do not see any instructions on how to run them. I believe that the authors should add them in the README or contribution guidelines. Additionally, they should consider automating the test execution with GitHub Actions (or a similar tool) to setup a git-ops CI workflow that requires tests to pass for each push/PR.
Reply: We have added instructions for running the tests in the README file. We have also updated the test functions and reduced download sizes to speed-up automated test (commits: 74e86796960ca5a6159bc8eac6a3b26ec730426a,
f2092a5ccd81ffacfa323557806c30930d492814,
6fc40847429b7580c8d01be936979a1e9a1a8e36,
07bc8e5c81fb6d71799a6781d6a812399ae66bb7).

### Software paper
- A statement of need: same as the point above in the "Documentation" section.
Reply: see reply above

- State of the field: again, in line with the first point of my review (in my other comment), I believe the authors should improve this part by commenting how geodata-harvester compares to other similar tools, e.g., intake.
Reply: See reply above. We added a comparison to other packages in the paper, including the intake package.

### Installation:
- A final note which I forgot: when I install geodata-harvester from conda-forge, I get a ModuleNotFoundError, which I do not get when I install it using pip.
Reply: We have not been able to reproduce this error. We have successfully tested multiple installations with conda and pip on different python environments. The latest geodata-harvester version is v1.1.0. Could you please let us know if you are still experiencing this issue?
