### 2023-06-26 #5205 @hugoledoux
> @sjsrey are you finding your way here? Have you found how it works?
>
> Could you please type this to ensure everything is working fine for you? `@editorialbot generate my checklist`

@sjsrey I see you generated the checklist, but since then nothing. Can you please give us a timeline? When can you make a first pass on the review? The 2 other reviewers have already commented a lot, this might help you too.
