### 2023-06-26 #5205 @sjsrey
Apologies for they delay.
I will get my review done by next monday.
Sergio (Serge) Rey
Professor
Department of Geography
Founding Director, Center for Open Geographical Science
San Diego State University

Schedule a meeting with me: https://calendly.com/sjsrey/office-hours

On Mon, Jun 26, 2023 at 5:26 AM Hugo Ledoux ***@***.***>
wrote:

> @sjsrey <https://github.com/sjsrey> are you finding your way here? Have
> you found how it works?
>
> Could you please type this to ensure everything is working fine for you? @editorialbot
> generate my checklist
>
> @sjsrey <https://github.com/sjsrey> I see you generated the checklist,
> but since then nothing. Can you please give us a timeline? When can you
> make a first pass on the review? The 2 other reviewers have already
> commented a lot, this might help you too.
>
