### 2023-07-17 #5205 @sjsrey
I am very sorry but I’ve had a family health crisis that has me out of commission. I’m going to have to withdraw from this and I apologize for not being able to help out this time.

Sergio (Serge) Rey
Professor
Department of Geography
Founding Director, Center for Open Geographical Science
San Diego State University
http://cogs.sdsu.edu

Schedule a meeting with me: https://calendly.com/sjsrey/office-hours

> On Jun 26, 2023, at 6:53 AM, Serge Rey ***@***.***> wrote:
>
> Apologies for they delay.
> I will get my review done by next monday.
> Sergio (Serge) Rey
> Professor
> Department of Geography
> Founding Director, Center for Open Geographical Science
> San Diego State University
>
> Schedule a meeting with me: https://calendly.com/sjsrey/office-hours
>
>
> On Mon, Jun 26, 2023 at 5:26 AM Hugo Ledoux ***@***.*** ***@***.***>> wrote:
>>
>> @sjsrey <https://github.com/sjsrey> are you finding your way here? Have you found how it works?
>>
>> Could you please type this to ensure everything is working fine for you? @editorialbot generate my checklist
>>
>> @sjsrey <https://github.com/sjsrey> I see you generated the checklist, but since then nothing. Can you please give us a timeline? When can you make a first pass on the review? The 2 other reviewers have already commented a lot, this might help you too.
>>
>> —
>> Reply to this email directly, view it on GitHub <https://github.com/openjournals/joss-reviews/issues/5205#issuecomment-1607367878>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAA42GXN4XIUERYBTWSFFG3XNF5YDANCNFSM6AAAAAAVM645BQ>.
>> You are receiving this because you were mentioned.
>>
