### 2023-07-26 #5205 @lukasbeuster
Hi @sebhaan,

I was indeed waiting for our third reviewer before commenting further (and was also on holiday for a bit), but here we go.

Thanks for your very detailed response to our review and all the work that you've put into addressing our comments. This is much appreciated!

Since you addressed the issues I raised very well I only have a few small things to note before I can tick the last items on the checklist:

1. Everything works as documented, but I am not getting results from the DEA dataset for the demo locations. All values in both ```DEA_ga_ls_ard_3_median_2021-09-29-to-2021-10-29``` and ```DEA_ga_ls_ard_3_median_2021-10-30-to-2021-11-29``` are ```-999```. What could be the reason for this? Is this replicable for you?

2. When going through the example_harvest_with_widget notebook, I encountered a little hiccup when importing an existing settings file using the widget. While trying to execute the harvest with ```df = harvest.run(fname_settings, return_df = True)```, I'm getting a ```NameError: name 'fname_settings' is not defined```

As far as I can deduct, this is due to the way the settings are loaded. When loading from the widget, you create fname_settings, but when you load from a file, this isn't done in the same way leading to errors as the file path isn't defined. Please let me know should this be simple user error :)

```
if w_load.value == None:
    dict_settings = hw.eval_widgets(w_settings, names_settings)
    # Convert settings from dictionary to SimpleNamespace (so all settings names available as settings.xxxname)
    settings = SimpleNamespace(**dict_settings)
    # remove GEE from settings if not used
    if settings.target_sources['GEE']['preprocess']['collection'] == None:
        del settings.target_sources['GEE']
    # Check if output path exists, if not create it:
    os.makedirs(settings.outpath, exist_ok=True)
    # Save settings to yaml file:
    fname_settings = os.path.join(settings.outpath, 'settings_saved.yaml')
    hw.save_dict_settings(dict_settings, fname_settings)
else:
    print(f'Settings loaded from {w_load.value}')
    settings = hw.load_settings(w_load.value)
# Print settings
hw.print_settings(settings)
```

Other than this I'm very happy with the submission and am looking forward to closing the checklist once I've received a response on the two points above.

Best
Lukas
