### 2023-08-21 ca8e820
Update notebook with latest code updates
