### 2023-08-21 dc6cd56
Update notebook with latest code updates
