### 2023-08-21 5e5d98b
Fix fname_settings in load_settings
