### 2023-08-22 #5205 @sebhaan
Hi @lukasbeuster,

thank you for pointing out these two issues, which have been addressed and fixed in the latest version (v1.1.1):

1) Problem in temporal aggregation of DEA data:
The temporal aggregation of DEA data layers using the median resulted in -999 values due to missing data and empty frames in the majority of the downloaded image for this example (this data source is published by DEA so we can't change that). Since the missing data value -999 seems to be not listed in the data file header of the corresponding DEA sources, we had to include additional checks for masking missing values before aggregation. This problem has been fixed in temporal.py by replacing these arbitrary missing data values with nans, which are then properly masked in the aggregation process (commit https://github.com/Sydney-Informatics-Hub/geodata-harvester/commit/8edf44c15d7db0fa614255d6e3dad051350d8078).

We have updated the corresponding example notebooks with the new processed data in `example_harvest.ipynb` and ``example_harvest_withGEE.ipynb` (commits https://github.com/Sydney-Informatics-Hub/geodata-harvester/commit/dc6cd56191d58a476ad8673ba2af6476ef92d5fb and https://github.com/Sydney-Informatics-Hub/geodata-harvester/commit/ca8e8204bcbe45e59d96cdb811ea5e77e22fe430)

2) Issue when importing an existing settings file using the widget notebook (example_harvest_with_widgets.ipynb).
This issue occurred while loading and executing a settings file via widgets. The issue has been easily fixed by adding  `fname_settings = w_load.value` (commit https://github.com/Sydney-Informatics-Hub/geodata-harvester/commit/5e5d98b1ef92e862be77905346f895750bb976f3).

The updates are included in the latest release notes for the minor update v1.1.1 (https://github.com/Sydney-Informatics-Hub/geodata-harvester/releases/tag/v1.1.1).

Once again, many thanks for your very valuable review and contribution to this package.

Best,
Seb
