### 2023-09-12 6b01b69
Some typos fixed + some suggestions
