### 2023-09-12 #5205 @hugoledoux
Okay, both reviewers are now happy, so at this point could you:

- [ ] Merge the PR I just made (https://github.com/Sydney-Informatics-Hub/geodata-harvester/pull/15) (yes I am a fan of the Oxford comma, and I improved I think at some places, feel free to reject if you prefer your words)
