### 2023-09-13 #5205 @sebhaan
>

Hi @hugoledoux ,

thank you for your work and paper edit suggestions.

- Your PR request with the revisions has been merged
- New proof generated and reviewed
- A new release (version tag v1.1.2) has been published: https://github.com/Sydney-Informatics-Hub/geodata-harvester/releases/tag/v1.1.2
- This latest release (v1.1.2) has been archived on Zenodo (DOI 10.5281/zenodo.8339817) with the correct metadata: https://zenodo.org/record/8339817

Thank you once again to both you and the reviewer for making this publication possible.
