+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Gorelick:2017,
   doi={10.1016/j.rse.2017.06.031},
-  title={Google Earth Engine: Planetary-scale geospatial analysis for everyone},
+  title={{Google Earth Engine}: {Planetary}-scale geospatial analysis for everyone},
   author={Gorelick, Noel and Hancher, Matt and Dixon, Mike and Ilyushchenko, Simon and Thau, David and Moore, Rebecca},
   journal={Remote sensing of Environment},
   volume={202},
@@ -11,7 +11,7 @@
 
 @article{Jeffrey:2001,
   doi={10.1016/S1364-8152(01)00008-1},  
-  title={Using spatial interpolation to construct a comprehensive archive of Australian climate data},
+  title={{Using} spatial interpolation to construct a comprehensive archive of {Australian} climate data},
   author={Jeffrey, Stephen J and Carter, John O and Moodie, Keith B and Beswick, Alan R},
   journal={Environmental Modelling \& Software},
   volume={16},
@@ -22,7 +22,7 @@
 }
 
 @article{Krause:2021,
-  title={Digital Earth Australia notebooks and tools repository},
+  title={{Digital Earth Australia} notebooks and tools repository},
   author={Krause, C and Dunn, B and Bishop-Taylor, R and Adams, C and Burton, C and Alger, M and Chua, S and Phillips, C and Newey, V and Kouzoubov, K and others},
   journal={Geoscience Australia},
   year={2021}
@@ -37,7 +37,7 @@
     number = {62},
     pages = {3168},
     author = {David Montero},
-    title = {eemont: A Python package that extends Google Earth Engine},
+    title = {eemont: {A Python} package that extends {Google Earth Engine}},
     journal = {Journal of Open Source Software}
 }
 
@@ -48,7 +48,7 @@
    volume = {XLVIII-4/W1-2022},
    pages = {301-306},
    author = {David Montero, Cesar Aybar, Miguel D. Mahecha, Sebastian Wieneke},
-   title = {spectral: Awesome Spectral Indices deployed via the Google Earth Engine JavaScript API},
+   title = {spectral: {Awesome} {Spectral} {Indices} deployed via the {Google Earth Engine JavaScript API}},
    journal = {The International Archives of the Photogrammetry, Remote Sensing and Spatial Information Sciences}
 }
 
@@ -61,5 +61,5 @@ volume = {5},
 number = {51}, 
 pages = {2305}, 
 author = {Qiusheng Wu}, 
-title = {geemap: A Python package for interactive mapping with Google Earth Engine}, 
+title = {geemap: {A Python} package for interactive mapping with {Google Earth Engine}}, 
 journal = {Journal of Open Source Software} }
