### 2023-09-14 #5205 @sebhaan
@kthyng
the capitalization in .bib has been corrected. Checked proof.
